Roctracer: use getApproximateTime() to avoid timestamp inversions with atem operator ranges
